Refactors Codebase after Code Review
