+++ pkg/dppapiservice/go/dpp_repository_service.go
@@ -32,11 +32,8 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
-	"io"
 	"net/http"
-	"net/url"
 	"sort"
-	"strconv"
 	"strings"
 	"time"
 
@@ -44,7 +41,6 @@ import (
 	"github.com/FriedJannik/aas-go-sdk/types"
 	aasrepositorydb "github.com/eclipse-basyx/basyx-go-components/internal/aasrepository/persistence"
 	submodelrepositorydb "github.com/eclipse-basyx/basyx-go-components/internal/submodelrepository/persistence"
-	"github.com/go-chi/chi/v5"
 )
 
 const (
@@ -594,261 +590,4 @@ func pageStrings(values []string, limit int32, cursor string) ([]string, string)
 		nextCursor = values[end-1]
 	}
 	return values[start:end], nextCursor
-}
-
-func errorResponse(status int, err error) ImplResponse {
-	return Response(status, Result{Messages: []Message{{
-		MessageType:   "Error",
-		Text:          err.Error(),
-		Code:          firstErrorCode(err.Error()),
-		CorrelationId: "",
-		Timestamp:     time.Now().UTC(),
-	}}})
-}
-
-func mapPersistenceError(err error, fallbackStatus int) ImplResponse {
-	status := fallbackStatus
-	text := strings.ToLower(err.Error())
-	if strings.Contains(text, "not found") || strings.Contains(text, "no rows") {
-		status = http.StatusNotFound
-	}
-	if strings.Contains(text, "duplicate") || strings.Contains(text, "already") {
-		status = http.StatusConflict
-	}
-	return errorResponse(status, err)
-}
-
-func firstErrorCode(text string) string {
-	fields := strings.Fields(text)
-	if len(fields) == 0 {
-		return "DPP-ERROR-UNKNOWN"
-	}
-	if strings.HasPrefix(fields[0], "DPP-") {
-		return fields[0]
-	}
-	return "DPP-ERROR-UNKNOWN"
-}
-
-// DPPRepositoryRouter exposes DPP repository service operations as HTTP handlers.
-type DPPRepositoryRouter struct {
-	service *DPPRepositoryService
-}
-
-// NewDPPRepositoryRouter creates an HTTP router adapter for the DPP repository service.
-func NewDPPRepositoryRouter(service *DPPRepositoryService) *DPPRepositoryRouter {
-	return &DPPRepositoryRouter{service: service}
-}
-
-// OrderedRoutes returns DPP routes in registration order.
-func (r *DPPRepositoryRouter) OrderedRoutes() []Route {
-	return []Route{
-		{"ReadDPPById", http.MethodGet, "/v1/dpps/{dppId}", r.ReadDPPById},
-		{"DeleteDPPById", http.MethodDelete, "/v1/dpps/{dppId}", r.DeleteDPPById},
-		{"UpdateDPPById", http.MethodPatch, "/v1/dpps/{dppId}", r.UpdateDPPById},
-		{"CreateDPP", http.MethodPost, "/v1/dpps", r.CreateDPP},
-		{"ReadDPPByProductId", http.MethodGet, "/v1/dppsByProductId/{productId}", r.ReadDPPByProductId},
-		{"ReadDPPVersionByIdAndDate", http.MethodGet, "/v1/dppsByIdAndDate/{dppId}", r.ReadDPPVersionByIdAndDate},
-		{"ReadDPPIdsByProductIds", http.MethodPost, "/v1/dppsByProductIds", r.ReadDPPIdsByProductIds},
-		{"ReadDataElement", http.MethodGet, "/v1/dpps/{dppId}/elements/*", r.ReadDataElement},
-		{"UpdateDataElement", http.MethodPut, "/v1/dpps/{dppId}/elements/*", r.UpdateDataElement},
-	}
-}
-
-// Routes returns DPP routes keyed by operation name.
-func (r *DPPRepositoryRouter) Routes() Routes {
-	routes := make(Routes)
-	for _, route := range r.OrderedRoutes() {
-		routes[route.Name] = route
-	}
-	return routes
-}
-
-// ReadDPPById handles GET /v1/dpps/{dppId}.
-func (r *DPPRepositoryRouter) ReadDPPById(w http.ResponseWriter, req *http.Request) {
-	representation, invalid := queryRepresentation(req)
-	if invalid != nil {
-		r.write(w, *invalid, nil)
-		return
-	}
-	response, err := r.service.ReadDPPById(req.Context(), pathParam(req, "dppId"), representation)
-	r.write(w, response, err)
-}
-
-// DeleteDPPById handles DELETE /v1/dpps/{dppId}.
-func (r *DPPRepositoryRouter) DeleteDPPById(w http.ResponseWriter, req *http.Request) {
-	response, err := r.service.DeleteDPPById(req.Context(), pathParam(req, "dppId"))
-	r.write(w, response, err)
-}
-
-// UpdateDPPById handles PATCH /v1/dpps/{dppId}.
-func (r *DPPRepositoryRouter) UpdateDPPById(w http.ResponseWriter, req *http.Request) {
-	body, err := readRequestBody(w, req)
-	if err != nil {
-		r.write(w, requestBodyErrorResponse("UPDDPP", err), nil)
-		return
-	}
-	response, err := r.service.UpdateDPPFromJSON(req.Context(), pathParam(req, "dppId"), body)
-	r.write(w, response, err)
-}
-
-// CreateDPP handles POST /v1/dpps.
-func (r *DPPRepositoryRouter) CreateDPP(w http.ResponseWriter, req *http.Request) {
-	body, err := readRequestBody(w, req)
-	if err != nil {
-		r.write(w, requestBodyErrorResponse("CREATEDPP", err), nil)
-		return
-	}
-	response, err := r.service.CreateDPPFromJSON(req.Context(), body)
-	r.write(w, response, err)
-}
-
-// ReadDPPByProductId handles GET /v1/dppsByProductId/{productId}.
-func (r *DPPRepositoryRouter) ReadDPPByProductId(w http.ResponseWriter, req *http.Request) {
-	representation, invalid := queryRepresentation(req)
-	if invalid != nil {
-		r.write(w, *invalid, nil)
-		return
-	}
-	response, err := r.service.ReadDPPByProductId(req.Context(), pathParam(req, "productId"), representation)
-	r.write(w, response, err)
-}
-
-// ReadDPPVersionByIdAndDate handles GET /v1/dppsByIdAndDate/{dppId}.
-func (r *DPPRepositoryRouter) ReadDPPVersionByIdAndDate(w http.ResponseWriter, req *http.Request) {
-	date, err := time.Parse(time.RFC3339Nano, req.URL.Query().Get("date"))
-	if err != nil {
-		r.write(w, errorResponse(http.StatusBadRequest, fmt.Errorf("DPP-READDATE-PARSE parse date query parameter: %w", err)), nil)
-		return
-	}
-	representation, invalid := queryRepresentation(req)
-	if invalid != nil {
-		r.write(w, *invalid, nil)
-		return
-	}
-	response, err := r.service.ReadDPPVersionByIdAndDate(req.Context(), pathParam(req, "dppId"), date, representation)
-	r.write(w, response, err)
-}
-
-// ReadDPPIdsByProductIds handles POST /v1/dppsByProductIds.
-func (r *DPPRepositoryRouter) ReadDPPIdsByProductIds(w http.ResponseWriter, req *http.Request) {
-	var request ReadDppIdsByProductIdsRequest
-	decoder := json.NewDecoder(http.MaxBytesReader(w, req.Body, maxDPPRequestBodyBytes))
-	decoder.DisallowUnknownFields()
-	if err := decoder.Decode(&request); err != nil {
-		r.write(w, requestBodyDecodeErrorResponse("READIDS", err), nil)
-		return
-	}
-	if err := validateReadDPPIdsRequest(request); err != nil {
-		r.write(w, errorResponse(http.StatusBadRequest, err), nil)
-		return
-	}
-	limit := defaultDPPPageLimit
-	if rawLimit := req.URL.Query().Get("limit"); rawLimit != "" {
-		parsed, err := strconv.ParseInt(rawLimit, 10, 32)
-		if err != nil || parsed < 1 {
-			r.write(w, errorResponse(http.StatusBadRequest, fmt.Errorf("DPP-READIDS-LIMIT invalid limit")), nil)
-			return
-		}
-		limit = int32(parsed)
-	}
-	response, err := r.service.ReadDPPIdsByProductIds(req.Context(), request, limit, req.URL.Query().Get("cursor"))
-	r.write(w, response, err)
-}
-
-func validateReadDPPIdsRequest(request ReadDppIdsByProductIdsRequest) error {
-	if len(request.ProductIds) == 0 {
-		return fmt.Errorf("DPP-READIDS-MISSING productIds must contain at least one product id")
-	}
-	if len(request.ProductIds) > maxDPPProductIDSearchItems {
-		return fmt.Errorf("DPP-READIDS-MAXITEMS productIds must contain at most %d product ids", maxDPPProductIDSearchItems)
-	}
-	for _, productID := range request.ProductIds {
-		if strings.TrimSpace(productID) == "" {
-			return fmt.Errorf("DPP-READIDS-INVALID productIds must contain only non-empty strings")
-		}
-	}
-	return nil
-}
-
-// ReadDataElement handles GET /v1/dpps/{dppId}/elements/{elementPath}.
-func (r *DPPRepositoryRouter) ReadDataElement(w http.ResponseWriter, req *http.Request) {
-	representation, invalid := queryRepresentation(req)
-	if invalid != nil {
-		r.write(w, *invalid, nil)
-		return
-	}
-	response, err := r.service.ReadDataElement(req.Context(), pathParam(req, "dppId"), elementPathParam(req), representation)
-	r.write(w, response, err)
-}
-
-// UpdateDataElement handles PUT /v1/dpps/{dppId}/elements/{elementPath}.
-func (r *DPPRepositoryRouter) UpdateDataElement(w http.ResponseWriter, req *http.Request) {
-	body, err := readRequestBody(w, req)
-	if err != nil {
-		r.write(w, requestBodyErrorResponse("UPDELEM", err), nil)
-		return
-	}
-	response, err := r.service.UpdateDataElementFromJSON(req.Context(), pathParam(req, "dppId"), elementPathParam(req), body)
-	r.write(w, response, err)
-}
-
-func readRequestBody(w http.ResponseWriter, req *http.Request) ([]byte, error) {
-	return io.ReadAll(http.MaxBytesReader(w, req.Body, maxDPPRequestBodyBytes))
-}
-
-func requestBodyErrorResponse(operation string, err error) ImplResponse {
-	if isRequestBodyTooLarge(err) {
-		return errorResponse(http.StatusRequestEntityTooLarge, fmt.Errorf("DPP-%s-BODYTOOLARGE request body exceeds %d bytes", operation, maxDPPRequestBodyBytes))
-	}
-	return errorResponse(http.StatusBadRequest, fmt.Errorf("DPP-%s-READBODY read request body: %w", operation, err))
-}
-
-func requestBodyDecodeErrorResponse(operation string, err error) ImplResponse {
-	if isRequestBodyTooLarge(err) {
-		return errorResponse(http.StatusRequestEntityTooLarge, fmt.Errorf("DPP-%s-BODYTOOLARGE request body exceeds %d bytes", operation, maxDPPRequestBodyBytes))
-	}
-	return errorResponse(http.StatusBadRequest, fmt.Errorf("DPP-%s-DECODE decode request body: %w", operation, err))
-}
-
-func isRequestBodyTooLarge(err error) bool {
-	return strings.Contains(err.Error(), "request body too large")
-}
-
-func pathParam(req *http.Request, name string) string {
-	return decodePathParam(chi.URLParam(req, name))
-}
-
-func elementPathParam(req *http.Request) string {
-	return strings.TrimPrefix(decodePathParam(chi.URLParam(req, "*")), "/")
-}
-
-func decodePathParam(value string) string {
-	decoded := value
-	for range 3 {
-		next, err := url.PathUnescape(decoded)
-		if err != nil || next == decoded {
-			return decoded
-		}
-		decoded = next
-	}
-	return decoded
-}
-
-func (r *DPPRepositoryRouter) write(w http.ResponseWriter, response ImplResponse, err error) {
-	if err != nil {
-		response = errorResponse(http.StatusInternalServerError, err)
-	}
-	_ = EncodeJSONResponse(response.Body, &response.Code, w)
-}
-
-func queryRepresentation(req *http.Request) (Representation, *ImplResponse) {
-	representation := Representation(req.URL.Query().Get("representation"))
-	if representation == "" {
-		return REPRESENTATION_COMPRESSED, nil
-	}
-	if !representation.IsValid() {
-		response := errorResponse(http.StatusBadRequest, fmt.Errorf("DPP-REPRESENTATION-INVALID invalid representation %q", representation))
-		return "", &response
-	}
-	return representation, nil
 }
